+++ lib/Exceptions/CreateWorkspaceException.php
@@ -1,10 +1,11 @@
+<?php
+
 /**
- * copyright Copyright (c) 2017 Arawa
+ * @copyright Copyright (c) 2017 Arawa
  *
- * author 2021 Baptiste Fotia <baptiste.fotia@arawa.fr>
- * author 2021 Cyrille Bollu <cyrille@bollu.be>
+ * @author 2022 Baptiste Fotia <baptiste.fotia@arawa.fr>
  *
- * license GNU AGPL version 3 or any later version
+ * @license GNU AGPL version 3 or any later version
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
@@ -21,7 +22,13 @@
  *
  */
 
-export const ESPACE_MANAGERS_PREFIX = 'GE-'
-export const ESPACE_USERS_PREFIX = 'U-'
-export const ESPACE_GID_PREFIX = 'SPACE-'
-export const PATTERN_CHECK_NOTHING_SPECIAL_CHARACTER = '[~<>{}|;.:,!?\'@#$+()%\\\\^=/&*[\\]]'
+namespace OCA\Workspace;
+
+class CreateWorkspaceException extends \Exception
+{
+
+	public function __construct($message, $code = 0)
+	{
+		parent::__construct($message, $code);
+	}
+}