+++ src/Home.vue
@@ -80,6 +80,12 @@
 					</div>
 				</div>
 			</template>
+			<ActionButton v-if="$root.$data.isUserGeneralAdmin === 'true'"
+				icon="icon-settings-dark"
+				class="btn-convert"
+				:close-after-click="true"
+				:title="t('workspace', 'Import / Convert')"
+				@click="toggleShowSelectGroupfoldersModal" />
 		</AppNavigation>
 		<AppContent>
 			<AppContentDetails>
@@ -93,11 +99,11 @@
 				</div>
 			</AppContentDetails>
 		</AppContent>
-		<Modal
+		<!-- <Modal
 			v-if="showSelectGroupfoldersModal"
 			@close="toggleShowSelectGroupfoldersModal">
 			<SelectGroupfolders @close="toggleShowSelectGroupfoldersModal" />
-		</Modal>
+		</Modal> -->
 	</Content>
 </template>
 
@@ -109,13 +115,14 @@ import AppNavigation from '@nextcloud/vue/dist/Components/AppNavigation'
 import AppNavigationIconBullet from '@nextcloud/vue/dist/Components/AppNavigationIconBullet'
 import AppNavigationItem from '@nextcloud/vue/dist/Components/AppNavigationItem'
 import AppNavigationNewItem from '@nextcloud/vue/dist/Components/AppNavigationNewItem'
-import Modal from '@nextcloud/vue/dist/Components/Modal'
-import ActionButton from '@nextcloud/vue/dist/Components/ActionButton'
 import Content from '@nextcloud/vue/dist/Components/Content'
 import { generateUrl } from '@nextcloud/router'
 import { getLocale } from '@nextcloud/l10n'
-import { get, formatGroups, create, formatUsers } from './services/groupfoldersService'
-import SelectGroupfolders from './SelectGroupfolders'
+import { get, formatGroups, createGroupfolder, formatUsers, checkGroupfolderNameExist, enableAcl, addGroupToGroupfolder, addGroupToManageACLForGroupfolder } from './services/groupfoldersService.js'
+import { createSpace, deleteBlankSpacename, isSpaceManagers, isSpaceUsers } from './services/spaceService.js'
+import { PATTERN_CHECK_NOTHING_SPECIAL_CHARACTER } from './constants.js'
+import NotificationError from './services/Notifications/NotificationError.js'
+import BadCreateError from './Errors/BadCreateError.js'
 
 export default {
 	name: 'Home',
@@ -125,15 +132,13 @@ export default {
 		AppNavigation,
 		AppNavigationIconBullet,
 		AppNavigationItem,
-		ActionButton,
 		AppNavigationNewItem,
 		Content,
-		Modal,
-		SelectGroupfolders,
 	},
 	data() {
 		return {
 			showSelectGroupfoldersModal: false,
+			// notificationError: NotificationError,
 		}
 	},
 	beforeCreate() {
@@ -265,64 +270,65 @@ export default {
 			}
 		},
 		// Creates a new space and navigates to its details page
-		createSpace(name) {
+		async createSpace(name) {
 			if (name === '') {
-				this.$notify({
+				const toastSpacenameEmpty = new NotificationError(this)
+				toastSpacenameEmpty.push({
 					title: t('workspace', 'Error'),
 					text: t('workspace', 'Please specify a name.'),
 					type: 'error',
 				})
 				return
 			}
-			const pattern = '[~<>{}|;.:,!?\'@#$+()%\\\\^=/&*[\\]]'
-			const regex = new RegExp(pattern)
-			if (regex.test(name)) {
-				this.$notify({
+			name = deleteBlankSpacename(name)
+
+			const REGEX_CHECK_NOTHING_SPECIAL_CHARACTER = new RegExp(PATTERN_CHECK_NOTHING_SPECIAL_CHARACTER)
+
+			if (REGEX_CHECK_NOTHING_SPECIAL_CHARACTER.test(name)) {
+				const toastCharacterNotAuthoized = new NotificationError(this)
+				toastCharacterNotAuthoized.push({
 					title: t('workspace', 'Error - Creating space'),
-					text: t('workspace', 'Your Workspace name must not contain the following characters: [ ~ < > { } | ; . : , ! ? \' @ # $ + ( ) % \\\\ ^ = / & * ]'),
+					text: t(
+						'workspace',
+						'Your Workspace name must not contain the following characters: [ ~ < > { } | ; . : , ! ? \' @ # $ + ( ) % \\\\ ^ = / & * ]',
+					),
 					duration: 6000,
-					type: 'error',
 				})
-				return
+				throw new BadCreateError(
+					'Your Workspace name must not contain the following characters: [ ~ < > { } | ; . : , ! ? \' @ # $ + ( ) % \\\\ ^ = / & * ]',
+				)
 			}
-			create(name)
-				.then(resp => {
-					if (resp.data.statuscode === 409) {
-						this.$notify({
-							title: t('workspace', 'Error - Creating space'),
-							text: t('workspace', 'This space or groupfolder already exist. Please, input another space.\nIf "toto" space exist, you cannot create the "tOTo" space.\nMake sure you the groupfolder doesn\'t exist.'),
-							type: 'error',
-						})
-					} else if (resp.data.statuscode === 400) {
-						this.$notify({
-							title: t('workspace', 'Error - Creating space'),
-							text: t('workspace', 'The groupfolder with this name : {spaceName} already exist', { spaceName: resp.data.spacename }),
-							duration: 6000,
-							type: 'error',
-						})
-					} else {
-						this.$store.commit('addSpace', {
-							color: resp.data.color,
-							groups: resp.data.groups,
-							isOpen: false,
-							id: resp.data.id_space,
-							groupfolderId: resp.data.folder_id,
-							name,
-							quota: t('workspace', 'unlimited'),
-							users: {},
-						})
-						this.$router.push({
-							path: `/workspace/${name}`,
-						})
-					}
-				})
-				.catch((e) => {
-					this.$notify({
-						title: t('workspace', 'Network error'),
-						text: t('workspace', 'A network error occured while trying to create the workspaces.') + '<br>' + t('workspace', 'The error is: ') + e,
-						type: 'error',
-					})
-				})
+
+			await checkGroupfolderNameExist(name, this)
+
+			const groupfolderId = await createGroupfolder(name, this)
+
+			await enableAcl(groupfolderId.data.id)
+
+			const workspace = await createSpace(name, groupfolderId.data.id, this)
+
+			const GROUPS_WORKSPACE = Object.keys(workspace.groups)
+			const workspaceManagerGid = GROUPS_WORKSPACE.find(isSpaceManagers)
+			const workspaceUserGid = GROUPS_WORKSPACE.find(isSpaceUsers)
+
+			await addGroupToGroupfolder(workspace.folder_id, workspaceManagerGid, this)
+			await addGroupToGroupfolder(workspace.folder_id, workspaceUserGid, this)
+
+			await addGroupToManageACLForGroupfolder(workspace.folder_id, workspaceManagerGid, this)
+
+			this.$store.commit('addSpace', {
+				color: workspace.color,
+				groups: workspace.groups,
+				isOpen: false,
+				id: workspace.id_space,
+				groupfolderId,
+				name,
+				quota: t('workspace', 'unlimited'),
+				users: {},
+			})
+			this.$router.push({
+				path: `/workspace/${name}`,
+			})
 		},
 		// Sorts groups alphabeticaly
 		sortedGroups(groups, space) {