+++ src/services/groupfoldersService.js
@@ -21,9 +21,17 @@
  *
  */
 
-import axios from '@nextcloud/axios'
+import { deleteBlankSpacename } from './spaceService.js'
 import { generateUrl } from '@nextcloud/router'
-import { createSpace, isSpaceManagers, isSpaceUsers } from './spaceService'
+import AddGroupToGroupfolderError from '../Errors/Groupfolders/AddGroupToGroupfolderError.js'
+import AddGroupToManageACLForGroupfolderError from '../Errors/Groupfolders/AddGroupToManageACLForGroupfolderError.js'
+import axios from '@nextcloud/axios'
+import BadGetError from '../Errors/BadGetError.js'
+import CheckGroupfolderNameExistError from '../Errors/Groupfolders/CheckGroupfolderNameError.js'
+import CreateGroupfolderError from '../Errors/Groupfolders/BadCreateError.js'
+import EnableAclGroupfolderError from '../Errors/Groupfolders/EnableAclGroupfolderError.js'
+import GetGroupfolderError from '../Errors/Groupfolders/GetGroupfolderError.js'
+import NotificationError from './Notifications/NotificationError.js'
 
 /**
  * @return {object}
@@ -35,8 +43,8 @@ export function getAll() {
 				return resp.data.ocs.data
 			}
 		})
-		.catch(e => {
-			console.error('Error to get all spaces', e)
+		.catch(error => {
+			throw new BadGetError('Error to get all spaces', error.reason)
 		})
 	return data
 }
@@ -44,26 +52,34 @@ export function getAll() {
 /**
  *
  * @param {number} groupfolderId it's the id of a groupfolder
- * @return {object}
+ * @param {object} vueInstance it's an instance of vue
+ * @return {Promise}
+ * @throws {GetGroupfolderError}
  */
-export function get(groupfolderId) {
-	const data = axios.get(generateUrl(`/apps/groupfolders/folders/${groupfolderId}`))
+export function get(groupfolderId, vueInstance = undefined) {
+	return axios.get(generateUrl(`/apps/groupfolders/folders/${groupfolderId}`))
 		.then(resp => {
 			if (resp.data.ocs.meta.status === 'ok') {
 				const workspace = resp.data.ocs.data
 				return workspace
+			} else {
+				throw new GetGroupfolderError('Impossible to get the groupfolder. May be an error network ?')
 			}
 		})
-		.catch(e => {
-			console.error('Error to get one space', e)
+		.catch((error) => {
+			const toastErrorToGetGroupfolder = new NotificationError(vueInstance)
+			toastErrorToGetGroupfolder.push({
+				title: t('workspace', 'Error to get the groupfolder'),
+				text: t('workspace', error.message),
+			})
+			throw new Error(error.message)
 		})
-	return data
 }
 
 /**
  *
  * @param {object} space it's an object relative to space
- * @return {object}
+ * @return {Promise}
  */
 export function formatGroups(space) {
 	const data = axios.post(generateUrl('/apps/workspace/api/workspace/formatGroups'), { workspace: space })
@@ -79,7 +95,7 @@ export function formatGroups(space) {
 /**
  *
  * @param {object} space it's an object relative to space
- * @return {object}
+ * @return {Promise}
  */
 export function formatUsers(space) {
 	const data = axios.post(generateUrl('/apps/workspace/api/workspace/formatUsers'), { workspace: space })
@@ -94,31 +110,44 @@ export function formatUsers(space) {
 
 /**
  * @param {string} spaceName it's the name of space to check
- * @return {boolean} true if such a groupfolder exists, false otherwise
+ * @param {object} vueInstance it's the instance of vue
+ * @return {Promise}
+ * @throws {CheckGroupfolderNameExistError}
  */
-function checkGroupfolderNameExist(spaceName) {
-	const groupfolders = getAll()
+export function checkGroupfolderNameExist(spaceName, vueInstance = undefined) {
+	return getAll()
 		.then(groupfolders => {
 			for (const folderId in groupfolders) {
 				if (spaceName.toLowerCase() === groupfolders[folderId].mount_point.toLowerCase()) {
-					return true
+					throw new CheckGroupfolderNameExistError('The groupfolder with this name : ' + spaceName + ' already exist')
 				}
 			}
-			return false
 		})
-		.catch(error => {
-			console.error('Cannot get all Groupfolders', error)
-			return false
+		.catch((error) => {
+			if (typeof (vueInstance) !== 'undefined') {
+				const toastSpaceOrGroupfoldersExisting = new NotificationError(vueInstance)
+				toastSpaceOrGroupfoldersExisting.push({
+					title: t('workspace', 'Error - Creating space'),
+					text: t(
+						'workspace',
+						'This space or groupfolder already exist. Please, input another space.'
+						+ '\nIf "toto" space exist, you cannot create the "tOTo" space.'
+						+ '\nMake sure you the groupfolder doesn\'t exist.',
+					),
+					duration: 6000,
+				})
+			}
+			throw new CheckGroupfolderNameExistError(error)
 		})
-	return groupfolders
 }
 
 /**
  * @param {number} folderId from a groupfolder
- * @return {object}
+ * @return {Promise}
+ * @throws {EnableAclGroupfolderError}
  */
 export function enableAcl(folderId) {
-	const result = axios.post(generateUrl(`/apps/groupfolders/folders/${folderId}/acl`),
+	return axios.post(generateUrl(`/apps/groupfolders/folders/${folderId}/acl`),
 		{
 			acl: 1,
 		})
@@ -126,20 +155,25 @@ export function enableAcl(folderId) {
 			if (resp.status === 200 && resp.data.ocs.meta.status === 'ok') {
 				return resp.data.ocs.data
 			}
+
+			if (resp.status === 500) {
+				throw new Error('Groupfolders\' API doesn\'t enable ACL. May be a problem with the connection ?')
+			}
 		})
 		.catch(error => {
-			console.error('Groupfolders\' API doesn\'t enable ACL. May be a problem with the connection ?', error)
+			throw new EnableAclGroupfolderError(error.message)
 		})
-	return result
 }
 
 /**
  * @param {number} folderId of an groupfolder
  * @param {string} gid it's an id (string format) of a group
- * @return {object}
+ * @param {object} vueInstance it's an instance of vue
+ * @return {Promise}
+ * @throws {AddGroupToGroupfolderError}
  */
-export function addGroup(folderId, gid) {
-	const result = axios.post(generateUrl(`/apps/groupfolders/folders/${folderId}/groups`),
+export function addGroupToGroupfolder(folderId, gid, vueInstance = undefined) {
+	return axios.post(generateUrl(`/apps/groupfolders/folders/${folderId}/groups`),
 		{
 			group: gid,
 		})
@@ -147,142 +181,90 @@ export function addGroup(folderId, gid) {
 			return resp.data.ocs.data
 		})
 		.catch(error => {
+			if (typeof typeof (vueInstance) !== 'undefined') {
+				const toastErrorToAddGroupToGroupfolder = new NotificationError(vueInstance)
+				toastErrorToAddGroupToGroupfolder.push({
+					title: t('workspace', 'Error groups'),
+					text: t('workspace', 'Impossible to attach the {error} group to groupfolder. May be a problem with the connection ?', { error }),
+				})
+			}
 			console.error(`Impossible to attach the ${gid} group to groupfolder. May be a problem with the connection ?`, error)
+			throw new AddGroupToGroupfolderError('Error to add Space Manager group in the groupfolder')
 		})
-	return result
 }
 
 /**
  * @param {number} folderId it's an id of a groupfolder
  * @param {string} gid it's an id (string format) of a group
- * @param {boolean} manageAcl (default: true)
- * @return {object} it's an object to check if it's a success or not
+ * @param {object} vueInstance it's an instance of vue
+ * @return {Promise}
+ * @throws {AddGroupToManageACLForGroupfolderError}
  */
-export function manageACL(folderId, gid, manageAcl = true) {
-	const result = axios.post(generateUrl(`/apps/groupfolders/folders/${folderId}/manageACL`),
+export function addGroupToManageACLForGroupfolder(folderId, gid, vueInstance) {
+	return axios.post(generateUrl(`/apps/groupfolders/folders/${folderId}/manageACL`),
 		{
 			mappingType: 'group',
 			mappingId: gid,
-			manageAcl,
+			manageAcl: true,
 		})
 		.then(resp => {
 			return resp.data.ocs.data
 		})
 		.catch(error => {
+			if (typeof (vueInstance) !== 'undefined') {
+				const toastErrorToAddGroupToManageACLForGroupfolder = new NotificationError(vueInstance)
+				toastErrorToAddGroupToManageACLForGroupfolder.push({
+					title: t('workspace', 'Error to add group as manager acl'),
+					text: t('workspace', 'Impossible to add the Space Manager group in Manage ACL groupfolder'),
+				})
+			}
 			console.error('Impossible to add the Space Manager group in Manage ACL groupfolder', error)
+			throw new AddGroupToManageACLForGroupfolderError('Error to add the Space Manager group in manage ACL groupfolder')
 		})
-	return result
 }
 
 /**
  * @param {string} spaceName it's the name space to create
- * @return {object} data
+ * @param {object} vueInstance it's the instance of vue
+ * @return {Promise}
+ * @throws {CreateGroupfolderError}
  */
-export async function create(spaceName) {
-	const data = { }
-	data.data = { }
-	const groupfolderIsExist = await checkGroupfolderNameExist(spaceName)
-	// Todo: Should I return a boolean value ?
-	if (groupfolderIsExist) {
-		data.data.message = `The groupfolder with this name : "${spaceName}" already exist`
-		data.data.statuscode = 400
-		data.data.spacename = spaceName
-		return data
-	}
-
-	// Create groupfolder
-	const groupfolderId = await axios.post(generateUrl('/apps/groupfolders/folders'),
+export function createGroupfolder(spaceName, vueInstance = undefined) {
+	return axios.post(generateUrl('/apps/groupfolders/folders'),
 		{
 			mountpoint: spaceName,
 		})
 		.then(resp => {
+			if (resp.data.ocs.meta.statuscode !== 100) {
+				throw new Error('Impossible to create a groupfolder. May be an error network ?')
+			}
 			return resp.data.ocs
 		})
 		.catch(error => {
-			console.error('Impossible to create a groupfolder. May be an error network ?', error)
-			data.data.statuscode = 400
-			return data
+			if (error instanceof Error) {
+				if (typeof (vueInstance) !== 'undefined') {
+					const toastErrorCreateGroupfolder = new NotificationError(vueInstance)
+					toastErrorCreateGroupfolder.push({
+						title: t('workspace', 'Error to create'),
+						text: t('workspace', error.message),
+					})
+				}
+				throw new Error(error.message)
+			}
+			if (typeof (vueInstance) !== 'undefined') {
+				const toastErrorNetworking = new NotificationError(vueInstance)
+				toastErrorNetworking.push({
+					title: t('workspace', 'Network error'),
+					text: t('workspace', 'A network error occured while trying to create the workspaces.'),
+				})
+			}
+			throw new CreateGroupfolderError('Network error - the error is: ' + error)
 		})
-	if (groupfolderId.meta.statuscode !== 100) {
-		console.error('Error when creating on groupfolder')
-		data.data.statuscode = 500
-		return data
-	}
-
-	// Get groupfolder created
-	const groupfolder = await get(groupfolderId.data.id)
-		.then(resp => {
-			return resp
-		})
-		.catch(error => {
-			console.error('Impossible to get the groupfolder. May be an error network ?', error)
-			data.data.statuscode = 400
-			return data
-		})
-
-	// Enable ACL on the groupfolder created
-	const aclIsEnabled = await enableAcl(groupfolder.id)
-	if (!aclIsEnabled.success) {
-		console.error('Error to enable acl')
-		data.data.statuscode = 500
-		return data
-	}
-	// Create the space
-	const resultCreateSpace = await createSpace(groupfolder.mount_point, groupfolder.id)
-	if (typeof (resultCreateSpace) !== 'object') {
-		console.error('Error when creating a space, it\'s not an object type.')
-		data.data.statuscode = 500
-		return data
-	}
-	// resultCreateSpace fill data
-	data.data = resultCreateSpace
-
-	// acl fill data
-	data.data.acl = {}
-	data.data.acl.state = true
-
-	// Add groups to groupfolder
-	const GROUPS = Object.keys(resultCreateSpace.groups)
-	const spaceManagerGID = GROUPS.find(isSpaceManagers)
-	const spaceUserGID = GROUPS.find(isSpaceUsers)
-
-	const isAddGroupForSpaceManager = await addGroup(resultCreateSpace.folder_id, spaceManagerGID)
-	if (!isAddGroupForSpaceManager.success) {
-		console.error('Error to add Space Manager group in the groupfolder')
-		data.data.statuscode = 500
-		return data
-	}
-
-	const isAddGroupForSpaceUser = await addGroup(resultCreateSpace.folder_id, spaceUserGID)
-	if (!isAddGroupForSpaceUser.success) {
-		console.error('Error to add Space Users group in the groupfolder')
-		data.data.statuscode = 500
-		return data
-	}
-
-	// Add Space Manager group in manage ACL
-	const resultManageACL = await manageACL(resultCreateSpace.folder_id, spaceManagerGID)
-	if (!resultManageACL.success) {
-		console.error('Error to add the Space Manager group in manage ACL')
-		console.error('GroupFolder API to manage ACL a groupfolder doesn\'t respond')
-		data.data.statuscode = 500
-		return data
-	}
-	// resultManageACL fill data
-	data.data.space_advanced_permissions = true
-	data.data.assign_permission = {
-		status: 'enabled',
-		groups: [
-			spaceManagerGID,
-		],
-	}
-
-	return data
 }
 
 /**
  * @param {object} workspace it's an object relative to workspace
- * @return {object}
+ * @return {Promise}
  */
 export function destroy(workspace) {
 	// It's possible to send data with the DELETE verb adding `data` key word as
@@ -315,7 +297,7 @@ export function destroy(workspace) {
  *
  * @param {object} workspace it's the object relative to workspace
  * @param {string} newSpaceName it's the new name for the workspace
- * @return {object}
+ * @return {Promise}
  */
 export function rename(workspace, newSpaceName) {
 	// Response format to return
@@ -331,6 +313,7 @@ export function rename(workspace, newSpaceName) {
 		console.error('The groupfolder name already exist. Please, choose another name to rename your space.')
 		return respFormat
 	}
+	newSpaceName = deleteBlankSpacename(newSpaceName)
 	// Update space side
 	const workspaceUpdated = axios.patch(generateUrl('/apps/workspace/api/space/rename'),
 		{